update roles v2 waypoint
