+++ packages/modules/package.json
@@ -24,7 +24,8 @@
     "ethers": "6.13.5",
     "evm-proxy-detection": "^2.0.1",
     "ser-kit": "1.1.0",
-    "zod": "^3.23.8"
+    "zod": "^3.23.8",
+    "zodiac-roles-deployments": "^2.2.2"
   },
   "devDependencies": {
     "@zodiac/eslint-config": "workspace:*",
